data visualization
Tableau Visualization:
Analyzed large dataset about 10,000 observations of the US Supermarket to recognize the revenue depending on various factors. Utilized Open Refine Tool extensively to sanitize the data and applied different visual models to forecast the revenue for upcoming years using Tableau. Created final report to suggest the range of products and location of stores help in increasing the revenue.